refactor build command into two files
